+++ src/lib.rs
@@ -46,17 +46,18 @@ pub mod netns;
 #[cfg(feature = "seccomp")]
 pub mod seccomp;
 
-#[cfg(feature = "dmverity")]
-pub mod dmverity;
-#[cfg(feature = "ima")]
-pub mod ima;
-// #[cfg(feature = "tpm")] pub mod tpm;
 #[cfg(feature = "certpin")]
 pub mod certpin;
+#[cfg(feature = "dmverity")]
+pub mod dmverity;
 #[cfg(feature = "fuse")]
 pub mod fuse;
+#[cfg(feature = "ima")]
+pub mod ima;
 #[cfg(feature = "pam")]
 pub mod pam;
+#[cfg(feature = "tpm")]
+pub mod tpm;
 
 #[cfg(feature = "audit")]
 pub mod audit;
@@ -64,11 +65,12 @@ pub mod audit;
 pub mod mac;
 // #[cfg(feature = "journald")] pub mod journald;
 // #[cfg(feature = "bootloader")] pub mod bootloader;
-// #[cfg(feature = "secureboot")] pub mod secureboot;
 #[cfg(feature = "agent")]
 pub mod agent;
 #[cfg(feature = "drm")]
 pub mod drm;
+#[cfg(feature = "secureboot")]
+pub mod secureboot;
 #[cfg(feature = "update")]
 pub mod update;